# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "AMONG API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3065",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/community/{communityId}/checkStatus": {
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "커뮤니티 등급 확인",
        "description": "커뮤니티 등급 확인",
        "parameters": [
          {
            "name": "communityId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/community/images": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "게시물 사진 저장",
        "description": "게시물 사진 저장",
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "description": "프로필 사진 주소"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/community/{communityId}/post": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "게시물 생성",
        "description": "게시물 생성",
        "parameters": [
          {
            "name": "communityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Post",
            "in": "body",
            "description": "게시물 생성",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/community/{postId}/post": {
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "게시물 상세 조회",
        "description": "게시물 상세 조회",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/community/{communityId}/{communityStatus}/{lastId}/posts": {
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "게시물 조회",
        "description": "게시물 조회",
        "parameters": [
          {
            "name": "communityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "communityStatus",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/community/{communityId}/{lastId}/posts": {
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "인플루언서 게시물 조회",
        "description": "인플루언서 게시물 조회",
        "parameters": [
          {
            "name": "communityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/community/post/{Postid}": {
      "patch": {
        "tags": [
          "Community"
        ],
        "summary": "게시물 수정",
        "description": "게시물 수정",
        "parameters": [
          {
            "name": "Postid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "post",
            "in": "body",
            "description": "게시물 예",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "내용 입력"
                },
                "image": {
                  "type": "string",
                  "example": "example.png"
                }
              },
              "required": [
                "content",
                "image"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/community/{postId}": {
      "delete": {
        "tags": [
          "Community"
        ],
        "summary": "게시물 삭제",
        "description": "게시물 삭제",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/community/{postId}/comment": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "댓글 생성",
        "description": "댓글 생성",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/community/{refId}/{PostId}/refcomment": {
      "post": {
        "tags": [
          "Community"
        ],
        "summary": "대댓글 생성",
        "description": "대댓글 생성",
        "parameters": [
          {
            "name": "refId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "PostId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/community/{postId}/like": {
      "patch": {
        "tags": [
          "Community"
        ],
        "summary": "좋아요",
        "description": "좋아요",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Community"
        ],
        "summary": "좋아요 삭제",
        "description": "좋아요 삭제",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/community/comment/{CommentId}": {
      "patch": {
        "tags": [
          "Community"
        ],
        "summary": "댓글 수정",
        "description": "댓글 수정",
        "parameters": [
          {
            "name": "CommentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/community/comment/{Commentid}": {
      "delete": {
        "tags": [
          "Community"
        ],
        "summary": "댓글 삭제",
        "description": "댓글 삭제",
        "parameters": [
          {
            "name": "Commentid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Post": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "example": "게시물 내용 입력"
        },
        "image": {
          "type": "array",
          "example": [
            "이미지 주소 입력",
            ""
          ],
          "items": {
            "type": "string"
          }
        },
        "limitedReader": {
          "type": "string",
          "example": "VIP"
        }
      }
    }
  }
}